chore(apps/web): Refactoring
