+++ packages/ui/src/GenericInputForm/AbiFields.tsx
@@ -1,77 +1,20 @@
 import { useFormContext } from "./context";
 import {
-    Alert,
-    Box,
-    Combobox,
     Flex,
     Group,
-    Input,
-    InputBase,
     SegmentedControl,
     Select,
     Stack,
     Text,
     Textarea,
-    TextInput,
     Tooltip,
     useCombobox,
 } from "@mantine/core";
-import { AbiFunction, AbiParameter } from "viem";
-import { FC, Fragment, useCallback, useEffect } from "react";
-import {
-    AbiValueParameter,
-    FormAbiMethod,
-    FormSpecification,
-    SpecificationMode,
-} from "./types";
-import { TbAlertCircle, TbHelp } from "react-icons/tb";
-import { encodeFunctionParams } from "./utils";
-import { useDebouncedCallback } from "@mantine/hooks";
-import LabelWithTooltip from "web/src/components/labelWithTooltip";
-import { AbiFunctionNameCombobox } from "./AbiFunctionNameCombobox";
-
-const placeholder = `function balanceOf(address owner) view returns (uint256) \nevent Transfer(address indexed from, address indexed to, uint256 amount)`;
-
-interface FunctionParamLabelProps {
-    input: AbiParameter;
-}
-
-export const FunctionParamLabel: FC<FunctionParamLabelProps> = ({ input }) => {
-    return (
-        <Box display="inline">
-            <Text c="cyan" span fz="sm">
-                {input.type}
-            </Text>{" "}
-            <Text span fz="sm">
-                {input.name}
-            </Text>
-        </Box>
-    );
-};
-
-interface FunctionSignatureProps {
-    abiFunction: AbiFunction;
-}
-
-export const FunctionSignature: FC<FunctionSignatureProps> = ({
-    abiFunction,
-}) => {
-    return (
-        <>
-            <Text span fw="bold" fz="sm">
-                {abiFunction.name}
-            </Text>
-            (
-            {abiFunction.inputs.map((input, index) => (
-                <Fragment key={`${input.type}-${input.name}`}>
-                    <FunctionParamLabel input={input} />
-                    {index < abiFunction.inputs.length - 1 ? ", " : ""}
-                </Fragment>
-            ))}
-            )
-        </>
-    );
-};
+import { FC } from "react";
+import { FormAbiMethod, FormSpecification, SpecificationMode } from "./types";
+import { TbHelp } from "react-icons/tb";
+import { AbiFunctionName } from "./AbiFunctionName";
+import { AbiFunctionParams } from "./AbiFunctionParams";
 
 export interface AbiFieldsProps {
     specifications: FormSpecification[];
@@ -79,10 +22,7 @@ export interface AbiFieldsProps {
 
 export const AbiFields: FC<AbiFieldsProps> = ({ specifications }) => {
     const form = useFormContext();
-    const { abiMethod, abiFunction, specificationMode, selectedSpecification } =
-        form.getTransformedValues();
-    const abiFunctionParams = form.getInputProps("abiFunctionParams");
-    const isFormValid = form.isValid();
+    const { abiMethod, specificationMode } = form.getTransformedValues();
     const specificationOptions = specifications.map((s) => ({
         value: s.id,
         label: s.name,
@@ -90,48 +30,6 @@ export const AbiFields: FC<AbiFieldsProps> = ({ specifications }) => {
     const combobox = useCombobox({
         onDropdownClose: () => combobox.resetSelectedOption(),
     });
-
-    const onChangeAbiFunctionName = useCallback(
-        (abiFunctionName: string) => {
-            combobox.closeDropdown();
-
-            form.setFieldValue("abiFunctionName", abiFunctionName);
-
-            const nextAbiFunction = (
-                (selectedSpecification?.abi as AbiFunction[]) ?? []
-            ).find((abiFunction) => abiFunction.name === abiFunctionName);
-
-            if (nextAbiFunction) {
-                const emptyFunctionParams = nextAbiFunction.inputs.map(
-                    (input) => ({
-                        ...input,
-                        value: "",
-                    }),
-                );
-
-                form.setFieldValue("abiFunctionParams", emptyFunctionParams);
-            }
-        },
-        [combobox, form, selectedSpecification],
-    );
-
-    const encodeFunctionParamsDebounced = useDebouncedCallback(
-        (params: AbiValueParameter[]) => {
-            // Encode the function params
-            const payload = encodeFunctionParams(params);
-            // Set the encoded function params as value for hex field
-            form.setFieldValue("rawInput", payload);
-        },
-        400,
-    );
-
-    const onChangeHumanAbiDebounced = useDebouncedCallback(() => {}, 400);
-
-    useEffect(() => {
-        if (isFormValid) {
-            encodeFunctionParamsDebounced(abiFunctionParams.value);
-        }
-    }, [abiFunctionParams.value, isFormValid, encodeFunctionParamsDebounced]);
 
     return (
         <Stack>
@@ -185,7 +83,7 @@ export const AbiFields: FC<AbiFieldsProps> = ({ specifications }) => {
                             data-testid="json-abi-textarea"
                             resize="vertical"
                             description="Define signatures in Human readable format"
-                            placeholder={placeholder}
+                            placeholder="function balanceOf(address owner) view returns (uint256) event Transfer(address indexed from, address indexed to, uint256 amount)"
                             rows={5}
                             label={
                                 <Group justify="flex-start" gap="3">
@@ -230,46 +128,8 @@ export const AbiFields: FC<AbiFieldsProps> = ({ specifications }) => {
                 />
             )}
 
-            <AbiFunctionNameCombobox />
-
-            {abiFunction && (
-                <Stack>
-                    {abiFunction.inputs.length > 0 ? (
-                        <>
-                            {abiFunction.inputs.map((input, index) => (
-                                <TextInput
-                                    key={`${input.name}-${input.type}`}
-                                    label={
-                                        input.name && input.type ? (
-                                            <FunctionParamLabel input={input} />
-                                        ) : (
-                                            input.name || input.type
-                                        )
-                                    }
-                                    placeholder={`Enter ${input.type} value`}
-                                    withAsterisk
-                                    {...form.getInputProps(
-                                        `abiFunctionParams.${index}.value`,
-                                    )}
-                                />
-                            ))}
-                        </>
-                    ) : (
-                        <Alert
-                            variant="light"
-                            color="blue"
-                            icon={<TbAlertCircle />}
-                            data-testid="empty-inputs-argments-alert"
-                        >
-                            No input arguments defined for{" "}
-                            <Text span fz="sm" fw="bold">
-                                {abiFunction.name}()
-                            </Text>
-                            .
-                        </Alert>
-                    )}
-                </Stack>
-            )}
+            <AbiFunctionName />
+            <AbiFunctionParams />
         </Stack>
     );
 };
